### 2020-10-01 #2716 @whedon
[PRE REVIEW]: REGENS: an open source Python package for simulating realistic autosomal genotypes
**Submitting author:** @greggj2016 (<a href="http://orcid.org/0000-0002-2619-3440">John Gregg</a>)
**Repository:** <a href="https://github.com/EpistasisLab/regens" target ="_blank">https://github.com/EpistasisLab/regens</a>
**Version:** v0.1.0
