### 2020-10-03 #2716 @greggj2016
Yes, actually this is the first paper I submitted here. I wasn't sure how to submit my reformatted draft, so I had attempted to resubmit the reformatted draft and withdraw the previous submission. However, clicking the button to withdraw the previous submission did not seem to do anything after multiple attempts. I was then advised to wait for your response, so this is not a hiccup in your system (Unless you count the withdraw submission button not working. You might want to look at that.)
