### 2020-10-03 #2716 @Kevin-Mattheus-Moerman
@greggj2016 Okay thanks for clarifying. In the future you can make changes to your paper on your repository and then call `@whedon generate pdf` (on its own, without any other text) here, to update the paper.
