### 2020-10-03 #2716 @greggj2016
Is there a place i can read more about the full functionality of whedon, or do I not need to know anything else?
