### 2020-10-03 #2716 @Kevin-Mattheus-Moerman
> Is there a place i can read more about the full functionality of whedon, or do I not need to know anything else?

See here: https://joss.readthedocs.io/en/latest/

Let me know if you have any more questions.
