### 2020-10-12 #2743 @whedon
[REVIEW]: REGENS: an open source Python package for simulating realistic autosomal genotypes
**Submitting author:** @greggj2016 (<a href="http://orcid.org/0000-0002-2619-3440">John Gregg</a>)
**Repository:** <a href="https://github.com/EpistasisLab/regens" target ="_blank">https://github.com/EpistasisLab/regens</a>
**Version:** v0.2.0c
