### 2020-10-24 #2743 @raivivek
@fboehm I am still unable to edit the checklist. I noticed that something similar happened in another thread and `whedon` was used with the following command: `@whedon re-invite @libertyh as reviewer` — perhaps this could help. Sorry for the trouble.
