### 2020-10-26 #2743 @raivivek
Thanks @fboehm! No worries at all. It's working as expected now. On the same note, I have completed my first-round of reviews and have a few revisions for @greggj2016 to consider.

Overall, I think the software paper and documentation is well done. The code for the software can use some refactoring but it doesn't prohibit use in its current form. I have highlighted some of those specific suggestions below.

---
- [ ] Keep only one LICENSE file (LICENSE with the copyright holder and year information within that file)
- [ ] Please create a new GitHub release with the latest code

- [ ] Create an installation package (should be available to install through `pip`) or at the very least add a  `requirements.txt` file with dependencies and minimum versions of the packages required
- [ ] Community guidelines - `CONTRIBUTING.md` is missing
- [ ] Provide supplementary details on the benchmark comparison made in the paper and the code to reproduce that in the manuscript

**Documentation** & **Software paper**
- Both are adequate and I've updated corresponding items in the reviewer checklist

**Code**
- [ ] Specify Python version support — Python 3 only or works across Python versions? What specific versions of Python supported, if any
- [ ] Code/additional files for the manuscript is currently bundled in the software repository. I'd recommend separating those. Creating an installation package should help (recommended in next point)
- [ ] Python code should be run through a code formatter such as [black](https://github.com/psf/black) to remove redundancies, code duplication, and other issues
- [ ] Unit testing module provides no feedback on what tests it's running and their status

- I **strongly recommend** creating an installation package such that REGENS can be installed through `pip` (even if the authors do not publish it on PyPI). Currently, the core of it is in two files `regens.py` and `regens_library.py`, and it creates opportunities for the tool to break if the library file is unknowingly moved by the user. This will also allow one to separate necessary data for the software from the manuscript code/docs contained in the repository using `package_data` or `package_dir` directives
    - More specifically, this would involve something like:
        - Testing related code to `test/`
        - References and other supplementary data to `data/`
        - Manuscript and other related docs to `docs/` or `manuscript/`

---

Also, @fboehm Let me know if you'd like me to open a new issue in the target repository with these suggestions. I'm still learning my way around JOSS reviews! :-) Thanks.
