### 2020-10-26 #2743 @fboehm
@raivivek - Thanks so much for your hard work on this! In future reviews, you can feel free to either list concerns in the review (like you've done here) and/or to open issues in the package submission repository. What you've done here is perfectly sufficient. Should you wish to open issues in the package repository, please be sure to link to them in the review.

I expect that the authors will appreciate your formulation of the edits as checklists. Thank you for being so thorough!
