### 2020-10-27 #2743 @raivivek
@greggj2016 Great that you brought this up. So the way I tested this is that I simply executed the file `regens_testers.py`. This gave no output and no feedback. Because there are no instructions as to how to run the tests, I presumed this would be the way to go. So I'd say as a first-step, you could revise the README to include unit-testing instructions. I'd also like to highlight that this would be easier and more can be done in more integrated fashion if you were to create a Python package. Let me know if I've missed anything.
