### 2020-10-27 #2743 @greggj2016
Sorry for the confusion. The file regens_testers.py only contains the testing functions. It does not implement them, which is why you received no output. Each test is its own folder and should be run after changing your working directory to that folder.
