### 2020-10-29 #2743 @raivivek
Hi @greggj2016 — thanks for the explanation. It's great to know that there are amazing tests bundled with REGENS. If possible, I'd encourage you to create a wrapper script that runs all of these tests for the end-user and reports results in one output. For example, the output exit status of this script can be used to ascertain if the software is working as expected or not in pipelines. Also, you may wish to avoid any kind of hardcoded values such as `source activate..` unless your install script creates such an environment. Thanks! Looking forward to all the changes.
