### 2020-10-30 #2743 @dwinter
Hi @greggj2016 and @raivivek,

I am working my way through this review and will have more detailed comments soon. Regens looks like a useful and well-written piece of software, and as far as I can tell it works as it says it does. I think think the major room for improvement will be in software engineering approaches that will make it easier for others ti work with or contribute to regens (and for you to maintain in ti future!).To that end, I wanted to reiterate a few points @raivivek has made.

- It is not clear to someone coming to the repo for the first time what test are available and how they might be run. The unitests don't use a testing module, rather they run a script that prints the resuls of each test to screen. Ideally, the existing tests could be moved to an automated testing environment (`unittest`, `pytest`.`nose` etc). At the least, it will be important to document the correct ways to run the tets and provide a high-level script that calls the tests without bouncing around between subdirectories.

- Ideally, the code should be provided as an installation package (https://packaging.python.org/tutorials/packaging-projects/), which wuold make it easier to install across platforms and handle requirements. If not, then there should certainly be a `requirements.txt` file to allow users to easily install the required packages via pip or conda.

Will provide a more detailed review next week.
