### 2020-11-02 EpistasisLab/regens#11 @dwinter
Installation procedure / instructions
This relates to my JOSS review openjournals/joss-reviews#2743

Installation instructions should include minimum versions for required packages (and indeed python) where applicable, and should be provided as a `requirements.txt` file (https://pip.pypa.io/en/stable/user_guide/#requirements-files). Thse files make installation of the required packages via `pip` or `anaconda` very straighforward.

The JOSS review criteria says that "ideally" installation can be handled via "[handled with an automated package management solution.](https://joss.readthedocs.io/en/latest/review_criteria.html#installation-instructions). Creating a package that included `regenens.py` as an executable would make make installation of the code more straightforward. (Even if the package is not uploaded to something like pip, being able to install through this approach would be helpful).
