### 2020-11-02 EpistasisLab/regens#12 @dwinter
Provide a high-level testing script
This relates to my JOSS review openjournals/joss-reviews#2743

`regens` contains a number of tests, spread across a few different working directories. At present, there is no clear instruction on how to run each test. In the least, it would be helpful to have

   - [ ] A high-level script for calling tests without needing to move from one subdirectory to another (perhaps defaulting to running all tests)
   - [ ] A README section clearly explaining how to run tests, and what each test is checking

The JOSS review guide says that software is strongly encouraged to include an automated test site, and that a good package will include 'An automated test suite hooked up to an external service such as Travis-CI or similar'. It would be much easier to move to an automated testing service if the current tests were translated into a python testing environment (e.g. `pytest` ,`nose` etc etc). If you chose to move the code to a standard package format for #1, the tests could be included in that structure and run automatically when the package is built.
