### 2020-11-02 EpistasisLab/regens#13 @dwinter
Clarify recombination map files
As discussed in openjournals/joss-reviews#2743, in the README

- [ ] Clarify that human recombination maps are included with teh code, so many users will not need to worry about the file format for these
- [ ] Detail the format for users that will need to make a custom one
