### 2020-11-02 EpistasisLab/regens#14 @dwinter
Add contributing.md
As noted by @raivivek in openjournals/joss-reviews#2743

At present `contributing.md` is referenced in the README included in the repo
