### 2020-11-02 #2743 @dwinter
It turns out my more complete review is not so different from the comments provide above. happy to continute discussions here or on the regens repo.
## Summary

`regens` is a well-written piece of software that will help researchers simulate datasets for GWAS. This includes realistic genotypes (including the LD structure of a given population) and phenotypic data (matching a given inheritance pattern and/or epistatic structure). The manuscript clearly states the need for such programs in the development of GWAS algorithms. Though my own research is not directly related to GWAS (so my insight might not be directly relevant) it does seem that  `regens` will prove helpful for researchers producing simulated datasets to compare methods, or addressing the feasibility of GWAS in novel populations (at least where the LD structure is known). So I think it meets the JOSS requirements of "Substantial scholarly effort".

Overall, the program is well-written. The core functions are all included in a single library file. The functions are well-documented and appropriately modular. High-level documentation is given via the README for the repository and an `examples` directory that includes scripts to run particular analyses and the required input files.

The package contains a number of tests. These are not implemented in a standard unit testing environment. Rather tests are spread across subdirectories, and partly written in the library functions and partly in another file. Although this is an unusual arrangement, the tests do seem to cover a good portion of the functionality and test of the package for the correctness of results from particular inputs.

The accompanying paper clearly states the need for the paper. The cited paper by Shi et al (2018) suggests the approach taken here simulates realistic data and the 'Differentiating attributes' section clearly compares this package with others that perform similar tasks.

## Major suggestions

Although my assessment is positive overall, I think the code could be improved by adopting software engineering approaches that make is easy for the current authors and future contributors to further develop the code and other users to incoprorate `regens` in their own workflows. I have started issues at the `regens` repo, which we can hopefully use to focus discussion on these points:

As noted in EpistasisLab/regens#1, installation instructions could be improved a little. Ideally the code could be provided as an installation package.

As noted EpistasisLab/regens#2, the tests provided with code are spread around a number of directories and it is not clear to someone new to the code exactly how they should be run. The key purpose of tests is to make sure future changes to the code don't introduce errors in output, so making the test-running as frictionless as possible is important.

## Minor comments

- It is a bit difficult to asses the `contribution and authorship' checkbox on the review, as it appears the code was only moved under version control recently (meaning commits only reflect a few weeks of work on the project). I am sure  @greggj2016 was the major contributor to this code prior to it coming under version control, but it might be useful to have this confirmed.

- Is there a reason to have a default value for the '--in' argument  'empty', rather than making it a required argument without a default? (`ArgParse` has the nice property of printing out a usage if a required argument is not included, which can be a helpful error message).

- In the `Input` subsection of the README, it ways in recombination map is given as 'a folder with one gzipped tab separated dataframe per chromosome. Each dataframe contains genomic positions of intervals' boundaries and the intervals' recombination rates.'. I think the files are actually recombination maps with the pysical (bp) v genetic (cM) positions of sites? And each interval is defined by a single position. Are these the centres or right-edges of genomic intervals? It would be useful to do two things (also listed in EpistasisLab/regens#3)

  + Make it clear that relevant files are provided for a lot of human populations (so many users needn't worry about this)
  + Clearly define the required format in case anyone has non-human data they   would like to use in this way

- I have also opened an issue for the missing contributing.md file (EpistasisLab/regens#4)
