### 2020-11-02 EpistasisLab/regens#13 @greggj2016
Thank you for your detailed review! I'll start addressing both of your suggestions tomorrow.

Just a note about issue #13: your first comment is addressed under the README file's "Input " section, and your second comment is addressed under the "Simulate genotype data with custom recombination rate dataframes" section. It would help me if you could clarify how I might make those sections more visible or simpler to understand.
