### 2020-11-02 #2743 @dwinter
A'huh. This is partly an issue of a tired reader and partly the fact the README is quite long!

I think it's always good to write teh README in the 'inverted pyramid' style with the most important information at the top. As most readers won't need to know about this format, and the current description confused at least one reader it might be best to have the first section to say  words to the effect of
(a) regens requires a recombination map
(b) maps are provided for many human populations
(c) custom maps can be provided, the format required for these detalied in SECTION XXX below"
