### 2020-11-02 #2743 @greggj2016
Another note regarding your minor comment:

First, and this is important to understand the recombination rate map file, centimorgans are true distances. Not many sources seem to explain it well. The best definition I came across was "One cM is equal to a 1% chance that a marker at one genetic locus will be separated from a marker at another locus [within the interval being examined] due to crossing over, (https://www.medicinenet.com/script/main/art.asp?articlekey=2665). However, this is still misleading. A better definition is that 1 centimorgan is equivalent to saying that the expected number of recombination events in the observed segment is 0.01. This is an important distinction when considering multiple segments: If you have two non-overlapping segments that are 1cM each, then the recombination rate of their union is (1cM + 1cM = 2cM), which corresponds to additivity of expected values, You can see how this paper (https://www.ncbi.nlm.nih.gov/pmc/articles/PMC52322/pdf/pnas01067-0025.pdf) uses it at the bottom of table 4, and how this paper (https://www.nature.com/articles/nature09525) uses it at the bottom of table 1 (i.e. they both add the cM measurements for all chromosomes).

After considering the first point, the second point is that the cM column contains _cumulative_ recombination rates, which means that the ith cM value is the recombination rate of the interval in between the first bp position and the ith bp position.  Additionally, genomic bp positions are interval left boundaries, which is why the "cM" element at the first position is 0. The cumulative recombination rate is necessarily 0 at the very first position because an interval cannot be considered without two positions.

Let me know If I can clarify this point further, It has been suggested that most readers won't care about these details, but I can include them if you'd like.
