### 2020-11-02 #2743 @greggj2016
Also, I think I understand and agree with the majority of your comments. Note though that this is my first serious coding project, so I was wondering if you could clarify what you meant when you said "I think the code could be improved by adopting software engineering approaches that make is easy for the current authors and future contributors to further develop the code and other users to incoprorate regens in their own workflows."

What does that mean specifically? Also, I'm wondering if pip is necessarily the right tool because regens reads plink files as input and writes plink files as output. There is no standard output for regens to pass directly to another program, particularly since it simulates only one chromosome at a time. Simulating all chromosomes in the python environment would require roughly 10-20 fold more RAM. I don't think many people would want to work with that, though it has been suggested that I might do this at some point. Still, it might require more work than expected at a first glance to implement efficiently, so I'd like to put that on the back-burner.
