### 2020-11-02 #2743 @moorejh28
I am the funder and supervisor of the project. Please accept this comment as confirmation that Mr. John Gregg, under my supervision, conceived of the project, planned the project, performed all the programming for the project, performed all the testing and evaluation of the code, wrote the JOSS paper, and created the Github project.
