### 2020-11-02 #2743 @trang1618
>It is a bit difficult to assess the `contribution and authorship' checkbox on the review, as it appears the code was only moved under version control recently (meaning commits only reflect a few weeks of work on the project). I am sure @greggj2016 was the major contributor to this code prior to it coming under version control, but it might be useful to have this confirmed.

Hi @dwinter I'm also confirming that @greggj2016 is the main author of the project.
