### 2020-11-02 #2743 @dwinter
Hi @greggj2016,

>Also, I think I understand and agree with the majority of your comments. Note though that this is my first serious coding project, so I was wondering if you could clarify what you meant when you said "I think the code could be improved by adopting software engineering approaches that make is easy for the current authors and future contributors to further develop the code and other users to incoprorate regens in their own workflows."

>What does that mean specifically...

This was really a preface dor the two major points I make in my review. I was trying to make it clear that paractices (automated unit testing and packaging) we have both suggested are helpful for development, and not just another hurdle that we are trying to place in your way.

I _think_ that suggesting these approaches is in line with the JOSS guidelines and ethos. Having said that, I've not reviwed for JOSS before. With two reviews in, it might be a good time to get some feedback on the next steps from our editor @fboehm ?
