### 2020-11-02 #2743 @dwinter
>Also, I'm wondering if pip is necessarily the right tool because regens reads plink files as input and writes plink files as output. There is no standard output for regens to pass directly to another program, and pip seems to be for installing functions used by the python environment, so i don't understand what purpose this serves

I don't think you should bundle a plink release with `regens`, much less implement these steps in python. The idea would be to include all of the python code in a pip package (no necessarily hosted on pip) that handles installing all of the other package/modules that are required. Even if you don't expect others to use your library code, a package will make it was to install the `regens.py` script into a users  `$PATH`.

Happy to continue this discussion, though it might be best to concentrate it under the EpistasisLab/regens#11?
