### 2020-11-02 EpistasisLab/regens#13 @dwinter
Just extending this with some points from our discussion openjournals/joss-reviews#2743.

I think the most important thing is that users know you provide maps for a lot of human populations (the hapmap ones, I think?), and they can be called from the regens commandline easily.

The second thing is to make it clear to a user that they can provide custom maps, and how they could provide them. The README section you have already covers this, though I would consider moving the example input closer to the top of the section. If this format is defined elsewhere (it looks like the first and third column of a `shapeit` map, for instance) you could also link to the definition.
