### 2020-11-02 ea64ccd
first attempt at automated installation
