+++ regens.py
@@ -18,235 +18,237 @@ from regens_library import breakpoints_to_simulated_individuals
 from regens_library import write_bed_file
 from regens_library import simulate_phenotypes
 
-parser = argparse.ArgumentParser()
-parser.add_argument('--in', nargs = "?", type = str, action = "store", dest = "input_plink_filename_prefix", default = "empty", help = 'input plink filename prefix')
-parser.add_argument('--out', nargs = "?", type = str, action = "store", dest = "output_plink_filename_prefix", default = "simulated_whole_genomes", help = 'output plink filename prefix')
-parser.add_argument('--simulate_nbreakpoints', nargs = "?", type = int, action = "store", dest = "num_breakpoints_simulated", default = 4, help = 'If there are n breakpoints, then each simulated chromosome is comprised of n+1 chunks, where each chunk is from a randomly drawn input sample')
-parser.add_argument('--simulate_nsamples', nargs = "?", type = int, action = "store", dest = "num_samples_simulated", default = 1000, help = 'number of simulated')
-parser.add_argument('--phenotype', nargs = "?", type = str, action = "store", dest = "phenotype", default = -9)
-parser.add_argument('--mean_phenotype', nargs = "?", type = float, action = "store", dest = "mean_phenotype", default = None)
-parser.add_argument('--population_code', nargs = "?", type = str, action = "store", dest = "population_code", default = "empty")
-parser.add_argument('--human_genome_version', nargs = "?", type = str, action = "store", dest = "human_genome_version", default = "empty")
-parser.add_argument('--recombination_file_path_prefix', nargs = "?", type = str, action = "store", dest = "recombination_file_path_prefix", default = "empty")
-parser.add_argument('--noise', nargs = "?", type = float, action = "store", dest = "noise", default = 0)
-parser.add_argument('--population_ID', nargs = "?", type = str, action = "store", dest = "population_ID", default = '')
-parser.add_argument('--causal_SNP_IDs_path', nargs = "?", type = str, action = "store", dest = "causal_SNP_IDs_path", default = "empty", help = 'input plink filename prefix')
-parser.add_argument('--major_minor_assignments_path', nargs = "?", type = str, action = "store", dest = "major_minor_assignments_path", default = "standard", help = 'input plink filename prefix')
-parser.add_argument('--betas_path', nargs = "?", type = str, action = "store", dest = "betas_path", default = "empty", help = 'input plink filename prefix')
-parser.add_argument('--SNP_phenotype_map_path', nargs = "?", type = str, action = "store", dest = "SNP_phenotype_map_path", default = "standard", help = 'input plink filename prefix')
-parser.add_argument('--test_functionality', nargs = "?", type = str, action = "store", dest = "test_functionality", default = "no", help = 'input plink filename prefix')
+def main():
+    parser = argparse.ArgumentParser()
+    parser.add_argument('--in', nargs = "?", type = str, action = "store", dest = "input_plink_filename_prefix", default = "empty", help = 'input plink filename prefix')
+    parser.add_argument('--out', nargs = "?", type = str, action = "store", dest = "output_plink_filename_prefix", default = "simulated_whole_genomes", help = 'output plink filename prefix')
[…]
