+++ setup.py
@@ -1,0 +1,6 @@
+from setuptools import setup
+
+setup(
+    name = 'regens',
+    entry_points = {'console_scripts': ['regens = regens:main']}
+)
