### 2020-11-03 #2743 @fboehm
Thanks, @dwinter, @raivivek and @greggj2016 ! The discussions around the reviews seem to be highly constructive and helpful, I hope, to @greggj2016 and continued development of REGENS. The next steps, I think, are for @greggj2016 to make changes to the software to address the issues raised in the reviews. Ultimately, we need both reviewers to be satisfied with the submission to the point where they're comfortable checking all of the boxes in the above checklists.

Unlike many journals, our review process at JOSS is highly iterative. It's best, I think, when there's a great deal of dialogue between reviewers and authors, like you all have done above.

@greggj2016 - do you feel that you have sufficient feedback via reviews and Github issues to begin refining the software? When you feel that you've resolved the issues, please comment again on this thread so that @raivivek and @dwinter may examine the fixes. They'll then provide additional feedback, stating that they're satisfied or unsatisfied with the fixes. If they're unsatisfied, they'll also explain what remains to be done. Does that make sense to everyone? Thanks again for your work on this!!
