+++ setup.py
@@ -1,7 +1,7 @@
 from setuptools import setup
 
 setup(
-    python_requires = ">=3.6"
+    python_requires = ">=3.6",
     name = 'regens',
     entry_points = {'console_scripts': ['regens = regens:main']},
     install_requires = ["bed-reader==0.1.1",
