### 2020-11-09 cd7f104
add contributing.md, closes #14
