### 2020-11-09 6e9b186
add contributing.md, closes #14 (#16)
