### 2020-11-13 #2743 @greggj2016
Yes, there has been a bit of a delay, which I apologize for. It has taken a while to figure out how to properly implement automated installation because this is my first serious project. My current strategy is to upload the code portion of my algorithm to pypi and the recombination rate files to figshare.com. I'll write an installation function that downloads the recombination rate files from figshare.com the first time that somebody calls regens.py. Other project have also eaten into my time. I hope to be done with this by Monday, but I'll give you guys a concrete update by then regardless.
