### 2020-11-13 7634138
add contributing.md, closes #14 (#17)
Co-authored-by: lelaboratoire <grixor@gmail.com>
