+++ setup.py
@@ -6,6 +6,7 @@ setup(
     python_requires = ">=3.7",
     name = 'regens',
     entry_points = {'console_scripts': ['regens = regens:main']},
+    setup_requires=["numpy"],
     install_requires = ["bed-reader==0.1.1",
                         "numpy==1.16.2",
                         "pandas==1.1.2",
