### 2020-11-14 #2743 @greggj2016
Full disclosure: My laptop that I started making the automated installation on broke down recently. I had a "new" laptop for a while, but I wanted to use the old one for as long as possible, and it stopped working just recently. My new laptop is apparently missing BLAS/LAPACK libraries that my old laptop already had, and when I try to get around that using anaconda, whatever compiler it uses is also missing some piece. Note that this doesn't change the existing project code at all: just my attempt at automated installation. It will take a little longer than expected, but it'll get done eventually.
