### 2020-11-15 #2743 @dwinter
Hi @greggj2016 That sounds like a plan. Feel free to sketch out your proposed solution before you pour lots of time into coding it up.
