### 2020-11-16 #2743 @greggj2016
Just want to say so far so good. I'm gonna first make the user manually install miniconda and a specific version of numpy. Then my current setup.py and _init_.py files are configured so that, when I type "python setup.py  install", it automatically installs regens in a way that allows one to run it from any directory. The recombination map files have to be in your working directory, and if they're not, then regens.py downloads them to your working directory automatically. It turns out that it's important that people use python version 3.7.x and an earlier version of numpy specifically. Part of the issue I was having turned out to be from an unstable new version of numpy, and the older version was apparently incompatible with python 3.8.x.

I still gotta figure out the best way to upload regens to pypi (my current repository is too large to upload to pypi) , and then it should be just small changes from there. Barring some new unexpected problem, I expect to be done with this soon.
