### 2020-11-16 #2743 @greggj2016
Question: To ensure that pypi is connected to this particular repository given its 60MB size limit, I intend to move the majority of non-python file content to figshare.com. There will be new functions in regens.py that automatically download the relevant content to your working directory on first use. Does that sound reasonable?
