### 2020-11-17 #2743 @raivivek
@greggj2016 Looks like good progress. On your question regarding data/recombination files, I think the figshare way sounds reasonable. However, there are other options as well in case you are interested. First option is that you can host the data on GitHub itself and download tarballs/zip from there. Second option is that you can request package size limit on PyPI to be increased so that you can bundle the data with the package itself. The benefit of second option is that it'll reduce friction in terms of writing additional code to download data and keeping the versioning consistent. You can request the package size limit here — https://github.com/pypa/packaging-problems. For example, see [here](https://github.com/pypa/packaging-problems/issues/92) or [here](https://github.com/pypa/pypi-support/issues/224).
