### 2020-11-18 6189d18
removed files unrelated to publication or source code and made corresponding changes to regens.py file and pypi related files
