+++ _init_.py
@@ -1,1 +1,3 @@
 from .regens import regens
+from .regens import regens_library
+from .regens import regens_testers
