### 2020-11-18 97c803f
Address comments (#18)
* first attempt at automated installation

* setup.py attempt 2

* added comma

* python version enforcement attempt 1

* added comma

* minor change

* Update README.md

* next setup.py attempt

* minor change

* another setup iteration

* removed files unrelated to publication or source code and made corresponding changes to regens.py file and pypi related files
