### 2020-11-18 #2743 @greggj2016
For right now, the larger files have been uploaded and annotated in figshare. It's actually better for me because, when regens is called as a package from the installation directory, the code I wrote still references many of the relevant files from the working directory.  It might be a little bit nicer if all files are downloaded to and accessible from the installation directory, but I also feel like that can be an improvement for later.
