### 2020-11-18 #2743 @greggj2016
This means I still gotta correct the syntax of my examples in the readme and some other minor things
