+++ README.md
@@ -3,52 +3,52 @@
 REGENS (REcombinatory Genome ENumeration of Subpopulations) is an open source Python package :package: that simulates whole genomes from real genomic segments. 
 REGENS recombines these segments in a way that simulates completely new individuals while simultaneously preserving the input genomes' linkage disequilibrium (LD) pattern with extremely high fedility. REGENS can also simulate mono-allelic and epistatic single nucleotide variant (SNV) effects on a continuous or binary phenotype without perturbing the simulated LD pattern.
 
-## Input :inbox_tray:
-REGENS requires the following inputs:
-- YOU MUST PROVIDE: real genotype data formatted as a standard (bed, bim, fam) plink _fileset_, ideally containing a minimum of 80 unrelated individuals.
-- WE HAVE PROVIDED: a recombination map for every 1000 genomes population. 
+### :star2: IMPORTANT NOTICE (PLEASE READ) :star2:
 
-The provided recombination maps were created by the [pyrho algorithm](https://github.com/popgenmethods/pyrho) and modified by us to minimize required disk space. [Recombination maps between related populations are highly correlated](https://github.com/EpistasisLab/regens/blob/address_comments/README.md#technical-details-robot), so you could pair your input dataset with the recombination map of the most genetically similar 1000 genomes population, as is usually done for SNP imputation. If you wish to make your own recombination map, then it must be formatted as described [here](https://github.com/EpistasisLab/regens/blob/address_comments/README.md#simulate-genotype-data-with-custom-recombination-rate-dataframes-abacus).  
+REGENS's simulated genomes are comprised entirely of concatenated segments from the input dataset's real genomes. If your input genomes are not available for public use, then you may not be allowed to publicly release the simulated dataset. Please consult the institutions that provide you access to your input genotype dataset for more information about this matter.
 
 ## Installing REGENS :hammer_and_wrench:
 
-Follow these instructions to install regens:
+Follow these instructions to install REGENS:
 
-1. [Install conda](https://docs.conda.io/en/latest/miniconda.html) if you haven't already installed either Anaconda or Miniconda.
+1. [Install conda](https://docs.conda.io/en/latest/miniconda.html) if you haven't already installed either Anaconda or Miniconda
 2. Open your conda terminal. Type "Anaconda" or "Miniconda" into your search bar and open the terminal. It will look like this: ![picture_of_conda_terminal](images/conda_terminal.png)
 3. Click the "Anaconda Prompt" app (left) to open the black terminal (right). The terminal's top must say "Anaconda prompt"
 4. Enter ```conda create --name regens python=3.7``` in the terminal to create a new environment called regens with python version 3.7
 5. Enter ```conda activate regens``` in the terminal to enter your new environment. If that doesn't work, enter ```source activate regens```
 6. Once in your regens environment (repeat step 5 if you close and reopen the conda terminal), enter ```pip install regens```
 
-Your first use of regens will download several files into your working directory, and you may experience permissions issue with this step. If so, try the following remedies:
+REGENS will download files into your working directory. If you experience permissions issues with this step, try the following remedies:
 
-1. Right click the "Anaconda Prompt" app (left), then click ```run as administrator```. If this helps, then it may be important to reinstall conda in a different directory. 
-2. Your antivirus software might block a file that regens needs (Avast blocked Miniconda's python.exe for us). Try seeing if your antivirus software is blocking anything related to anaconda or regens, and then allow it to stop blocking that file. You could also turn off your antivirus software, though we do not recommend this.  
[…]
