+++ README.md
@@ -12,12 +12,12 @@ REGENS's simulated genomes are comprised entirely of concatenated segments from 
 Follow these instructions to install REGENS:
 
 1. [Install conda](https://docs.conda.io/en/latest/miniconda.html) if you haven't already installed either Anaconda or Miniconda
-2. Open your conda terminal. Type "Anaconda" or "Miniconda" into your search bar and open the terminal. It will look like this: ![picture_of_conda_terminal](images/conda_terminal.png)
+2. Open your conda terminal. Type "Anaconda" or "Miniconda" into your search bar and open the terminal. It will look like this: <img src="images/conda_terminal.png" width="800" height="275"/>
 3. Click the "Anaconda Prompt" app (left) to open the black terminal (right). The terminal's top must say "Anaconda prompt"
 4. Enter ```conda create --name regens python=3.7``` in the terminal to create a new environment called regens with python version 3.7
 5. Enter ```conda activate regens``` in the terminal to enter your new environment. If that doesn't work, enter ```source activate regens```
 6. Once in your regens environment (repeat step 5 if you close and reopen the conda terminal), enter ```pip install regens```
-7. Run [this command](https://github.com/EpistasisLab/regens/blob/final_changes/README.md#simulate-genotype-data-computer) to allow regens to download the remaining files. It will write the simulated data into the `examples` folder that it downloads. If you experience permissions issues with this step, [try these remedies]():
+7. Run [this command](https://github.com/EpistasisLab/regens/blob/final_changes/README.md#simulate-genotype-data-computer) to allow regens to download the remaining files. It will write the simulated data into the `examples` folder that it downloads. If you experience permissions issues with this step, [try these remedies](https://github.com/EpistasisLab/regens/blob/final_changes/README.md#remedies-to-known-permission-issues-adhesive_bandage):
 
 ## Input :turkey:
 REGENS requires the following inputs:
