### 2020-11-20 93fdb0f
Final changes (#19)
* Update README.md

* picture of what conda looks like

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* Update README.md
