+++ README.md
@@ -3,51 +3,39 @@
 REGENS (REcombinatory Genome ENumeration of Subpopulations) is an open source Python package :package: that simulates whole genomes from real genomic segments. 
 REGENS recombines these segments in a way that simulates completely new individuals while simultaneously preserving the input genomes' linkage disequilibrium (LD) pattern with extremely high fedility. REGENS can also simulate mono-allelic and epistatic single nucleotide variant (SNV) effects on a continuous or binary phenotype without perturbing the simulated LD pattern.
 
-## Input :inbox_tray:
+## :star2: IMPORTANT NOTICE (PLEASE READ) :star2:
+
+REGENS's simulated genomes are comprised entirely of concatenated segments from the input dataset's real genomes. If your input genomes are not available for public use, then you may not be allowed to publicly release the simulated dataset. Please consult the institutions that provide you access to your input genotype dataset for more information about this matter.
+
+## Instructions to Installing REGENS :hammer_and_wrench:
+
+1. [Install conda](https://docs.conda.io/en/latest/miniconda.html) if you haven't already installed either Anaconda or Miniconda
+2. Open your conda terminal. Type "Anaconda" or "Miniconda" into your search bar and open the terminal. It will look like this: <img src="images/conda_terminal.png" width="800" height="400"/>
+3. Click the "Anaconda Prompt" app (left) to open the black terminal (right). The terminal's top must say "Anaconda prompt"
+4. Enter ```conda create --name regens python=3.7``` in the terminal to create a new environment called regens with python version 3.7
+5. Enter ```conda activate regens``` in the terminal to enter your new environment. If that doesn't work, enter ```source activate regens```
+6. Once in your regens environment (repeat step 5 if you close and reopen the conda terminal), enter ```pip install regens```
+7. Run [this command](https://github.com/EpistasisLab/regens/blob/final_changes/README.md#simulate-genotype-data-computer) to allow regens to download the remaining files. It will write the simulated data into the `examples` folder that it downloads. If you experience permissions issues with this step, [try these remedies](https://github.com/EpistasisLab/regens/blob/final_changes/README.md#remedies-to-known-permission-issues-adhesive_bandage):
+
+## Input :turkey:
 REGENS requires the following inputs:
 - YOU MUST PROVIDE: real genotype data formatted as a standard (bed, bim, fam) plink _fileset_, ideally containing a minimum of 80 unrelated individuals.
 - WE HAVE PROVIDED: a recombination map for every 1000 genomes population. 
 
 The provided recombination maps were created by the [pyrho algorithm](https://github.com/popgenmethods/pyrho) and modified by us to minimize required disk space. [Recombination maps between related populations are highly correlated](https://github.com/EpistasisLab/regens/blob/address_comments/README.md#technical-details-robot), so you could pair your input dataset with the recombination map of the most genetically similar 1000 genomes population, as is usually done for SNP imputation. If you wish to make your own recombination map, then it must be formatted as described [here](https://github.com/EpistasisLab/regens/blob/address_comments/README.md#simulate-genotype-data-with-custom-recombination-rate-dataframes-abacus).  
 
-## Installing REGENS :hammer_and_wrench:
-
-Follow these instructions to install regens:
-
-1. [Install conda](https://docs.conda.io/en/latest/miniconda.html) if you haven't already installed either Anaconda or Miniconda.
-2. Open your conda terminal. Type "Anaconda" or "Miniconda" into your search bar and open the terminal. It will look like this: ![Real and simulated R value vs. MAF](images/r_maf_ACB.png)
-3. Enter ```conda create --name regens python=3.7``` to create a new environment called regens with python version 3.7.
[…]
