### 2020-11-20 9718e56
including thinning methods for 1000 genome SNPS\
