+++ thinning_methods/get_1000_genomes_files.sh
@@ -1,0 +1,590 @@
+#!/bin/bash
+#BSUB -J get_1000_genomes_files
+#BSUB -o get_1000_genomes_files.out
+#BSUB -e get_1000_genomes_files.error
+module load plink/1.90Beta
+module load plink/2.0
+
+#IMPORTANT: This file is not ready to run. It only lists the plink operations that were used to make the processed 1000 genomes datasets.
+
+# gets 1000 genome files
+# wget -O all_1000_genomes.pgen.zst https://www.dropbox.com/s/afvvf1e15gqzsqo/all_phase3.pgen.zst?dl=1
+# wget -O all_1000_genomes.pvar.zst https://www.dropbox.com/s/op9osq6luy3pjg8/all_phase3.pvar.zst?dl=1
+# wget -O all_1000_genomes.psam https://www.dropbox.com/s/nhfhskyy50sqsf1/phase3_orig.psam?dl=1
+
+# decompresses 1000 genome files
+plink2 --zst-decompress all_1000_genomes.pgen.zst > all_1000_genomes.pgen
+plink2 --zst-decompress all_1000_genomes.pvar.zst > all_1000_genomes.pvar
+
+# removes variants without at least 5 instances, without which, every european subpopulation cannot have an instance.
+plink2 --pfile all_1000_genomes --maf 0.002 --make-pgen --out all_1000_genomes_filtered
+rm all_1000_genomes.pgen
+rm all_1000_genomes.pvar
+
+# removes non-autosomes, filters multi-allelic variants, filters, and seperates european subpopulations.
+# This makes the all_1000_genomes_[POPULATION]_filter.txt file for every population.
+python all_1000_genomes_subpop_filter_getter.py
+plink2 --pfile all_1000_genomes_filtered --max-alleles 2 --chr 1-22 --keep all_1000_genomes_ACB_filter.txt --make-bed --out all_1000_genomes_ACB
+plink2 --pfile all_1000_genomes_filtered --max-alleles 2 --chr 1-22 --keep all_1000_genomes_ASW_filter.txt --make-bed --out all_1000_genomes_ASW
+plink2 --pfile all_1000_genomes_filtered --max-alleles 2 --chr 1-22 --keep all_1000_genomes_BEB_filter.txt --make-bed --out all_1000_genomes_BEB
+plink2 --pfile all_1000_genomes_filtered --max-alleles 2 --chr 1-22 --keep all_1000_genomes_CDX_filter.txt --make-bed --out all_1000_genomes_CDX
+plink2 --pfile all_1000_genomes_filtered --max-alleles 2 --chr 1-22 --keep all_1000_genomes_CEU_filter.txt --make-bed --out all_1000_genomes_CEU
+plink2 --pfile all_1000_genomes_filtered --max-alleles 2 --chr 1-22 --keep all_1000_genomes_CHB_filter.txt --make-bed --out all_1000_genomes_CHB
+plink2 --pfile all_1000_genomes_filtered --max-alleles 2 --chr 1-22 --keep all_1000_genomes_CHS_filter.txt --make-bed --out all_1000_genomes_CHS
+plink2 --pfile all_1000_genomes_filtered --max-alleles 2 --chr 1-22 --keep all_1000_genomes_CLM_filter.txt --make-bed --out all_1000_genomes_CLM
[…]
