### 2020-11-20 #2743 @greggj2016
Okay, I think it's almost ready.

1) I would like to know if you think that the new installation instructions and description of the recombination rate files in the README are satisfactory.
2) I will write a detailed document about all of the tests and a shell script that runs the other testing shell scripts later tonight. In the future, I will probably centralize this with github's continuous integration feature, though this is not a priority for me right now.
3) I think there might be one typo in the paper to fix.

After accounting for these three things, I would like to know if you guys think this package is ready for publication.
