### 2020-11-20 #2743 @greggj2016
I also have another more specific question: Can I leave the efficiency testing of Triadsim out of the master unit testing script? The problem is that I don't want to assume anything about the number of cores that the user has, and Triadsim takes a really long time to run and a huge amount of RAM. Triadsim also has to be installed separately, so I would rather leave that particular component to the user who wants to test that.
