+++ regens_tests_info.md
@@ -1,0 +1,7 @@
+There are four testing folders:
+1. correctness_testing_ACB
+2. correctness_testing_GBR
+3. runtime_testing
+4. unit_testing_files
+
+The tests in "correctness_testing_ACB" and "correctness_testing_GBR" ensure that the most important parts of the regens algorithm work correctly for two different samples. 
