### 2020-11-20 84eebe3
Update regens_tests_info.md
