### 2020-11-21 46526aa
Update regens_tests_info.md
