### 2020-11-21 6d64a9a
Update regens_tests_info.md
