### 2020-11-21 #2743 @greggj2016
Just another note: There was one line in my code:

model = LogisticRegression(C = 1E100, tol = 1E-100, max_iter = 1000000).fit(features, simulated_phenotypes)

failing to include solver = 'lbfgs' (even though that is the default value) caused this line to print a statement that, somehow, causes it to freeze when simulating phenotypes. Specifying that solver = 'lbfgs' in the function fixes this, but now I have to update pypi. I'm not sure what commands I need to update the package on pypi, so I'll probably be done with this tomorrow (possibly Monday) instead of today.
