### 2020-11-23 #2743 @greggj2016
Yes, we have decided to implement Github Actions in order to update the pypi software. This may take a little while longer, but it's really something we were going to do anyway.  It might as well be now.
