### 2020-11-24 EpistasisLab/regens#13 @greggj2016
I have addressed both comments in the following commit:

https://github.com/EpistasisLab/regens/pull/18/commits/bb75edf59b530050e2abfb9387dc9be4afa627c2

This commit is labeled `Update README.md` in this merge:

https://github.com/EpistasisLab/regens/pull/18
