### 2020-11-24 EpistasisLab/regens#13 @greggj2016
I put a link to the example input in the input description.

I also put a link to the input description in the input example:

https://github.com/EpistasisLab/regens/commit/fc8a4009b39e7fb5e1363a43603e1dd9c4e73b7b

The tables are output from the [pyrho algorithm](https://github.com/popgenmethods/pyrho), and I can't say with any certainty if they follow a particular format, but I do mention this algorithm a few times.
