### 2020-11-24 84bfe45
Update regens_tests_info.md
