+++ regens_tests_info.md
@@ -1,29 +1,19 @@
 ## Summary
 
-There is one file named `run_all_tests.sh` in REGENS' repository. If you run REGENS in an anaconda envrionment named `regens` (follow the installation instructions in the README), then running the command `bsub < run_all_tests.sh` will most of the tests that regens downloads. There are four testing folders: 
-
-WARNING: the runtime_testing files are not ready to run. They all contain the line `#BSUB -m lambda[i]`, where i is the node. Change this to your node names and make sure that all nodes that you use have identical processors. 
+There is one file named `run_all_tests.sh` in REGENS' repository. If you run REGENS in an anaconda envrionment named `regens` (follow the installation instructions in the README), then running the command `bsub < run_all_tests.sh` will run the majority of tests that regens downloads. There are four testing folders: 
 
 1. correctness_testing_ACB
 2. correctness_testing_GBR
 3. runtime_testing
 4. unit_testing_files
 
-REQUIREMENTS: running `run_all_tests.sh` submits 13 parallel jobs, each of which takes roughly 10GB of RAM (slight overestimate) and 5 minutes to complete (rough estimate). You you may want to run one job at a time, in which case, you should change the number of samples being generated so that you have enough RAM. You should also run the following files consecutively (paths are all relative to your working directory):
+WARNING: the runtime_testing files are not ready to run. They all contain the line `#BSUB -m lambda[i]`, where i is the node. Change this to your node names and make sure that all nodes that you use have identical processors. 
+
+REQUIREMENTS: running `run_all_tests.sh` submits 3 parallel jobs. The correctness_testing jobs take roughly 4.5GB of RAM 2.5 minutes each to complete. The unit_testing jobs take less than 1GB of RAM and 30 seconds to complete. If you want to run one job at a time, then you should run the following files consecutively (paths are all relative to your working directory):
 
 1. correctness_testing_ACB/regens_automated_tests_part1_ACB.sh
 2. correctness_testing_GBR/regens_automated_tests_part1_GBR.sh
 3. unit_testing_files/unit_tests.sh
-4. runtime_testing/regens1.sh
-5. runtime_testing/regens2.sh
-6. runtime_testing/regens3.sh
-7. runtime_testing/regens4.sh
-8. runtime_testing/regens5.sh
-9. runtime_testing/regens6.sh
-10. runtime_testing/regens7.sh
-11. runtime_testing/regens8.sh
-12. runtime_testing/regens9.sh
-13. runtime_testing/regens10.sh
 
 Note: all tests take a relatively long time to run, and in general, should not be ran if using regens to simulate a large quantity of data. Tests that `run_all_tests.sh` will run are described below. 
 
@@ -31,16 +21,18 @@ Note: all tests take a relatively long time to run, and in general, should not b
 
 The tests in `correctness_testing_ACB` and `correctness_testing_GBR` ensure that the most important parts of the regens algorithm work correctly for two different samples. Specifically, it tests four different things:
 
-1. For all possible breakpoint intervals (to which dataset SNPs are not yet assigned) that the expected number of breakpoints in each interval is, on average, equal to the observed number of breakpoints in each interval. It does this for every chromosome, and names the files for the ith chromosome `expected_vs_actual_breakpoint_counts_full_range_chr[i].png`. Files corresponding to the ACB and GBR profiles get written into the corresponding correctness_testing folders. 
[…]
