### 2020-11-24 6502a4e
makes standard testing simpler
