### 2020-11-24 EpistasisLab/regens#12 @greggj2016
commits related to the testing README are below:

https://github.com/EpistasisLab/regens/commit/4b5a4af2a8712ca61c61a700ca81151c99750f3f
https://github.com/EpistasisLab/regens/commit/84eebe3807bfaacd078f90597ec5be728782398d
https://github.com/EpistasisLab/regens/commit/46526aa4dcc8b9b81a392f5ec82c541ceea1d264
https://github.com/EpistasisLab/regens/commit/6d64a9a29f16d949611cbb18d6711d9e52c12f75
https://github.com/EpistasisLab/regens/commit/84bfe457ea44388186f30ee3f90063abb8adc29b
