### 2020-11-24 EpistasisLab/regens#12 @greggj2016
commits related to the high-level script are below:

https://github.com/EpistasisLab/regens/commit/6b5c15a367d32fe44beb51ba1481485de4b4c296
https://github.com/EpistasisLab/regens/commit/6502a4e3d3b606ee833993bc2529c09f9b83dc17 (this one is most recent)
