### 2020-11-24 EpistasisLab/regens#11 @greggj2016
The following pull request contains a number of commits related to setting up the automated installation

https://github.com/EpistasisLab/regens/pull/18

Within that are some of the more important commits:
https://github.com/EpistasisLab/regens/pull/18/commits/6189d18958b3516eb12e62f6ea07952c3fcb0821 shows removing files that are now downloaded from figshare.com.
https://github.com/EpistasisLab/regens/pull/18/commits/ea64ccddbdefb6e290c10202bb31b0a51c8b9cb0 shows modifying regens.py to be the main function for setuptools.
https://github.com/EpistasisLab/regens/pull/18/commits/c3d5d31d7f678f2921d2bbfc911aee87f2c68f51 shows an early incomplete version of setup.py
Other commits entail me trying things that mostly didn't work. They're less important.

This commit includes the majority of other changes to setup.py and changes to the yml file that automates pypi updates

https://github.com/EpistasisLab/regens/commit/8cfbe54c90767460a9827a68d03e59780287d727

The following pull request includes a number of commit hashes that contributed to writing the new installation instructions. I won't list them because I edited the readme directly, and they're not particularly organized.

https://github.com/EpistasisLab/regens/pull/19

The following commits entail changing some links that make the installation instructions a little bit easier to follow:

https://github.com/EpistasisLab/regens/commit/a29f4fc658f043874848567d63c81c5cb828f37e
https://github.com/EpistasisLab/regens/commit/8319813c5f91fc678d2029d18a679a64da8e3eef
https://github.com/EpistasisLab/regens/commit/94489e837214539d61ed92c12fd07fbaf154577e
https://github.com/EpistasisLab/regens/commit/8d17524f6f806a790ab90c62050a858b83c9a7f4
https://github.com/EpistasisLab/regens/commit/0dea489dfefa86014076d52f545056b94ef6bf97
https://github.com/EpistasisLab/regens/commit/6040a4d12666e556f8738a85474577a589dfabc9
https://github.com/EpistasisLab/regens/commit/9ebfac42ac5ec569874702674638c3287f01438d
