+++ setup.py
@@ -14,8 +14,8 @@ setup(
     long_description = regens_summary,
     long_description_content_type = 'text/markdown',
     packages = find_packages(),
-    version='0.2.0b',
-    python_requires = "==3.7",
+    version='0.2.0',
+    python_requires = "==3.7.*",
     name = 'regens',
     entry_points = {'console_scripts': ['regens = regens:main']},
     py_modules = ['regens', 'regens_library', 'regens_testers'],
