### 2020-11-24 EpistasisLab/regens#11 @greggj2016
These commits are related to updating the pypi to its final version:

https://github.com/EpistasisLab/regens/commit/087b9da37939d7782a229854c35bfbc3d1e8b5f2
https://github.com/EpistasisLab/regens/commit/b4a93097c7085c205b6dc36a0737f093a34a71fb

Please feel free to reopen the issue if these changes have not adequately addressed your concern!
