### 2020-11-24 #2743 @greggj2016
Actually, never mind, I may be missing one thing:

> * []  Provide supplementary details on the benchmark comparison made in the paper and the code to reproduce that in the manuscript

Can you please clarify what you're looking for here?
