### 2020-11-25 4de06b3
removing thining methods that are now in regens-analysis repository
