+++ regens_library.py
@@ -9,36 +9,38 @@ from copy import deepcopy as COPY
 from sklearn.linear_model import LogisticRegression
 from sklearn.linear_model import LinearRegression
 from regens_testers import unit_tester
+
 pd.options.mode.chained_assignment = None
 
-# TODO: double check that the modified recombination_rate_to_probabilities function works as expected. 
-# TODO: double check why you don't subtract 1 from the first "SNP_positions_to_rcmb_intervals" use but you do from the second. 
-# IMPORTANT DEFINITION: a breakpoint refers to the genomic position of the boundary between segments. 
-# from two sampled whole genomes. It does not refer to the programming definition of breakpoint.  
+# TODO: double check that the modified recombination_rate_to_probabilities function works as expected.
+# TODO: double check why you don't subtract 1 from the first "SNP_positions_to_rcmb_intervals" use but you do from the second.
+# IMPORTANT DEFINITION: a breakpoint refers to the genomic position of the boundary between segments.
+# from two sampled whole genomes. It does not refer to the programming definition of breakpoint.
+
 
 def reduce_recomb_rate_info(rcmb_rate_info, bim_SNP_positions, test_functionality):
 
     """
-    
+
     Purpose
     -------
-    To reduce the number of rows in rcmb_rate_info so that every resulting genomic interval contains at least one SNP. 
+    To reduce the number of rows in rcmb_rate_info so that every resulting genomic interval contains at least one SNP.
 
     Parameters
     ----------
     rcmb_rate_info: a pandas dataframe with two columns. "Position(bp)" is the genomic position of the ith interval's left
-                    boundary. The ith row is both the left boundry of the ith interval and the right boundary of the (i-1)th 
+                    boundary. The ith row is both the left boundry of the ith interval and the right boundary of the (i-1)th
                     interval. "Map(cM)" is the cumulative recombination rate in centiMorgans, which is 0 in the first row.
 
-    bim_SNP_positions: the genomic positions of every SNP directly from the input bim file. 
+    bim_SNP_positions: the genomic positions of every SNP directly from the input bim file.
 
     Returns
     -------
-    reduced_rcmb_rate_info: "rcmb_rate_info" with rows removed to ensure that at least one 
-                            SNP position from "bim_SNP_positions" resides in every interval. 
-    
+    reduced_rcmb_rate_info: "rcmb_rate_info" with rows removed to ensure that at least one
+                            SNP position from "bim_SNP_positions" resides in every interval.
+
     """
-    
+
     rcmb_rate_intervals = rcmb_rate_info["Position(bp)"].to_numpy()
 
     # INDEXING NOTE: Not subtracting one from "SNP_pos_rcmb_interval_map" returns the closest indices of "rcmb_rate_intervals"
@@ -46,25 +48,34 @@ def reduce_recomb_rate_info(rcmb_rate_info, bim_SNP_positions, test_functionalit
     #                The only row still needed is the closest boundary to the LEFT of the first SNP. This is aquired by
     #                implementing "occupied_rcmb_intervals[np.min(np.where(occupied_rcmb_intervals == True)) - 1] = True".
 
-    SNP_pos_rcmb_interval_map = SNP_positions_to_rcmb_intervals(rcmb_rate_intervals, COPY(bim_SNP_positions), test_functionality, context = 1)
+    SNP_pos_rcmb_interval_map = SNP_positions_to_rcmb_intervals(
+        rcmb_rate_intervals, COPY(bim_SNP_positions), test_functionality, context=1
+    )
     all_rcmb_intervals = np.arange(len(rcmb_rate_intervals))
     occupied_rcmb_intervals = np.isin(all_rcmb_intervals, SNP_pos_rcmb_interval_map)
[…]
