### 2020-11-25 #2743 @greggj2016
In combination with responding to the issues raised on our repository, we believe we have done all of the following:

> **General checks**
>
> * [x]  Keep only one LICENSE file (LICENSE with the copyright holder and year information within that file)
> * [x]  Please create a new GitHub release with the latest code
>
> **Functionality**
>
> * [x]  Create an installation package (should be available to install through `pip`) or at the very least add a  `requirements.txt` file with dependencies and minimum versions of the packages required
> * [x]  Community guidelines - `CONTRIBUTING.md` is missing
> * [x]  Provide supplementary details on the benchmark comparison made in the paper and the code to reproduce that in the manuscript
>
> **Documentation** & **Software paper**
>
> * Both are adequate and I've updated corresponding items in the reviewer checklist
>
> **Code**
>
> * [x]  Specify Python version support — Python 3 only or works across Python versions? What specific versions of Python supported, if any
> * [x]  Please add package version information in the tool. For example, `regens.py --version` should tell the current version
> * [x]  Code/additional files for the manuscript is currently bundled in the software repository. I'd recommend separating those. Creating an installation package should help (please see the end of this review)
> * [x]  Python code should be run through a code formatter such as [black](https://github.com/psf/black) to remove redundancies, code duplication, and other issues
> * [x]  Unit testing module provides no feedback on what tests it's running and their status
> * [x]  Create an installation package such that REGENS can be installed through pip
