### 2020-11-25 #2743 @greggj2016
We have also made and submitted a small number of minor changes to the paper. We are also still updating the supplementary repository (https://github.com/EpistasisLab/regens-analysis), though it has details on our benchmark tests and instructions on how to rerun them (instructions are [here](https://github.com/EpistasisLab/regens/blob/main/regens_tests_info.md) in the main repository)
