### 2021-03-04 #2743 @dwinter
Let me start my sending apologies to @greggj2016  and co-authors and to @fboehm about taking so long to get to these revisions. If I could have predicted how busy I would be now when I accepted this review I wouldn't have taken this on!

I have now finally found the time to look at the revised version of regens closely, and I am happy to say the authors have met all of my concerns and comments. I think the code now meets the expectations of the JOSS review process.

My only small suggestion would be to include a link to the `regens_tests_info.md` file in the README (possibly in the repository structure section). Just to provide a would-be developer with a path to the files and some explanation of what each file is testing.

I hope you go something from our reviews, well done on meeting them so well, and sorry again for taking so long to sign off on the revised code.
