### 2021-03-04 #2743 @greggj2016
Thank you for your guidance throughout this review! No worries about your busy spell, it happens to us all! We have addressed your final suggested edit [here](https://github.com/EpistasisLab/regens/commit/14b64e4fb27afd5b3cbbdcf51570ed6fc8ebe6ff).

Big thanks to @fboehm also for your facilitation and patience! It’s been a great experience, and I have learned a lot!
