### 2021-03-04 #2743 @fboehm
@greggj2016 - It looks like the spacing in the YAML of paper.md is off a little. I'd like for the pdf to have the affiliations/ addresses on separate lines. Right now, affiliation "2" starts right after the end of affiliation 1, ie, on the same line. Let's see if we can fix that.

Can you try adding two blank spaces after each line in the YAML? I'm not sure if that will fix the pdf rendering, but it might. Thanks again!!
