+++ paper/paper.md
@@ -5,6 +5,7 @@ tags:
   - data simulation
   - bioinformatics
   - python
+
 authors:
   - name: John T. Gregg  
     orcid: 0000-0002-2619-3440  
@@ -15,11 +16,14 @@ authors:
   - name: Jason H. Moore  
     orcid: 0000-0002-5015-1099  
     affiliation: 2  
+    
 affiliations:  
   - name: Department of Biostatistics, Epidemiology and Informatics, University of Pennsylvania, Philadelphia, PA 19087, USA  
     index: 1  
+    
   - name: Institute for Biomedical Informatics, University of Pennsylvania, Philadelphia, PA 19087, USA  
     index: 2  
+    
 date: 01 October 2020  
 bibliography: paper.bib  
 ---
