### 2021-03-04 #2743 @trang1618
> Can you try adding two blank spaces after each line in the YAML? I'm not sure if that will fix the pdf rendering, but it might. Thanks again!!

Hi @fboehm thank you so much for your help! I've tried adding different spaces in this [PR](https://github.com/EpistasisLab/regens/pull/20) but still didn't fix the issue ([Whedon preview page](https://whedon.theoj.org/) gave the same pdf for this branch). I also so see a few recent JOSS papers that have this format (e.g. [this]( https://joss.theoj.org/papers/10.21105/joss.02957 ) and [this](https://joss.theoj.org/papers/10.21105/joss.02781)), so I'm wondering if this is a new change? In any case, if you would like to try some different changes in the PR, please do not hesitate! Thanks a lot!
