### 2021-03-04 #2743 @danielskatz
@greggj2016 - I just found one small change in proof-reading the paper, as shown in https://github.com/EpistasisLab/regens/pull/21  If you can merge this, we can move to the final processing and publishing.
